Update library Telegram Bot
Вынес UpdateHandler в отдельный класс.
