+++ TelegramBot/TiskTask.TelegramBot/TelegramBot.cs
@@ -1,4 +1,8 @@
-﻿using Telegram.Bot;
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+using Telegram.Bot;
 using Telegram.Bot.Polling;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
@@ -8,6 +12,9 @@ namespace TiskTask.TelegramBot
 {
     public class TelegramBot
     {
+        /// <summary>
+        /// Создание клиента для работы с Телеграм ботом.
+        /// </summary>
         private readonly ITelegramBotClient _botClient;
 
         public TelegramBot(string botToken)
@@ -20,76 +27,18 @@ namespace TiskTask.TelegramBot
         /// </summary>
         public async Task StartAsync(CancellationToken cancellationToken = default)
         {
+            /// <summary>
+            /// Запрос информации о боте.
+            /// </summary>
             var me = await _botClient.GetMe();
-            Console.WriteLine($"✅ Бот @{me.Username} запущен. Ожидание сообщений...");
+
+            Console.WriteLine($"Бот @{me.Username} запущен. Ожидание сообщений...");
 
             // Запускаем получение обновлений
             await _botClient.ReceiveAsync(
                 updateHandler: new UpdateHandler(_botClient),
                 cancellationToken: cancellationToken
             );
-        }
-    }
-
-    /// <summary>
-    /// Обработчик входящих обновлений от Telegram.
-    /// </summary>
-    internal class UpdateHandler : IUpdateHandler
-    {
-        private readonly ITelegramBotClient _botClient;
-
-        public UpdateHandler(ITelegramBotClient botClient)
-        {
-            _botClient = botClient;
-        }
-
-        public async Task HandleUpdateAsync(ITelegramBotClient botClient, Update update, CancellationToken cancellationToken)
-        {
-            try
-            {
-                if (update.Message is not Message message) return;
-
-                var chatId = message.Chat.Id;
-                var text = message.Text;
-
-                Console.WriteLine($"📩 От {message.From.FirstName}: {text ?? "[не текст]"}");
-
-                if (text == "/start")
-                {
-                    await botClient.SendMessage(
-                        chatId: chatId,
-                        text: "Добро пожаловать!",
-                        cancellationToken: cancellationToken
-                    );
-                    return;
-                }
-
-                if (message.Type == MessageType.Text && !string.IsNullOrEmpty(text))
-                {
-                    await botClient.SendMessage(
-                        chatId: chatId,
-                        text: $"📝 Вы написали: {text}",
-                        cancellationToken: cancellationToken
-                    );
-                }
-                else
-                {
-                    await botClient.SendMessage(
-                        chatId: chatId,
-                        text: "Я могу обрабатывать только текстовые сообщения.",
-                        cancellationToken: cancellationToken
-                    );
-                }
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine($"❌ Ошибка при обработке сообщения: {ex.Message}");
-            }
-        }
-
-        public Task HandleErrorAsync(ITelegramBotClient botClient, Exception exception, HandleErrorSource source, CancellationToken cancellationToken)
-        {
-            throw new NotImplementedException();
         }
     }
 }
