+++ TelegramBot/TiskTask.TelegramBot/UpdateHandler.cs
@@ -1,0 +1,85 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
+using Telegram.Bot.Polling;
+using Telegram.Bot.Types.Enums;
+using Telegram.Bot.Types;
+using Telegram.Bot;
+
+namespace TiskTask.TelegramBot
+{
+    /// <summary>
+    /// Обработчик входящих обновлений от Telegram.
+    /// </summary>
+    internal class UpdateHandler : IUpdateHandler
+    {
+        /// <summary>
+        /// Создание клиента для работы с Телеграм ботом.
+        /// </summary>
+        private readonly ITelegramBotClient _botClient;
+
+        public UpdateHandler(ITelegramBotClient botClient)
+        {
+            _botClient = botClient;
+        }
+
+        public async Task HandleUpdateAsync(ITelegramBotClient botClient, Update update, CancellationToken cancellationToken)
+        {
+            try
+            {
+                if (update.Message is not Message message) return;
+
+                /// <summary>
+                /// Получение Id пользователя.
+                /// </summary>
+                var chatId = message.Chat.Id;
+
+                /// <summary>
+                /// Получение текста от пользователя.
+                /// </summary>
+                var text = message.Text;
+
+                Console.WriteLine($"📩 От {message.From.FirstName}: {text ?? "[не текст]"}");
+
+                if (text == "/start")
+                {
+                    await botClient.SendMessage(
+                        chatId: chatId,
+                        text: "Добро пожаловать!",
+                        cancellationToken: cancellationToken
+                    );
+                    return;
+                }
+
+                if (message.Type == MessageType.Text && !string.IsNullOrEmpty(text))
+                {
+                    await botClient.SendMessage(
+                        chatId: chatId,
+                        text: $"📝 Вы написали: {text}",
+                        cancellationToken: cancellationToken
+                    );
+                }
+                else
+                {
+                    await botClient.SendMessage(
+                        chatId: chatId,
+                        text: "Я могу обрабатывать только текстовые сообщения.",
+                        cancellationToken: cancellationToken
+                    );
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine($"❌ Ошибка при обработке сообщения: {ex.Message}");
+            }
+        }
+
+        public Task HandleErrorAsync(ITelegramBotClient botClient, Exception exception, HandleErrorSource source, CancellationToken cancellationToken)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
